--- lib/bin.js
@@ -1,6 +1,0 @@
-#!/usr/bin/env node
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-require("chalk");
-console.log("MAML; ");
-//# sourceMappingURL=data:application/json;base64,eyJ2ZXJzaW9uIjozLCJmaWxlIjoiYmluLmpzIiwic291cmNlUm9vdCI6IiIsInNvdXJjZXMiOlsiLi4vc3JjL2Jpbi50cyJdLCJuYW1lcyI6W10sIm1hcHBpbmdzIjoiOzs7QUFDQSxpQkFBZTtBQUNmLE9BQU8sQ0FBQyxHQUFHLENBQ1AsUUFBUSxDQUNYLENBQUMifQ==
