+++ lib/cli/bin.d.ts
