+++ package.json
@@ -25,12 +25,12 @@
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
   "scripts": {
-    "build": "rimraf lib && tslint \"{src,test}/**/*.ts\" && tsc --pretty --inlineSourceMap && chmod a+x ./lib/hc.js ./lib/index.js",
+    "build": "rimraf lib && tslint \"{src,test}/**/*.ts\" && tsc --pretty --inlineSourceMap && npm run set-bin",
     "clean": "rimraf lib && rm -f *.log.*",
     "debug": "echo 'Use `debugger;`\n' && inspect _mocha --no-timeouts --compilers ts:ts-node/register --recursive test/**/*-spec.ts",
     "lint": "tslint --force --format verbose \"{src,test}/**/*.ts\"",
-    "run": "node lib/hc.js",
-    "set-bin": "chmod a+x ./lib/hc.js ./lib/index.js",
+    "hc": "node lib/cli/hc.js",
+    "set-bin": "chmod a+x ./lib/cli/hc.js ./lib/cli/bin.js",
     "test": "npm run build && mocha --require ts-node/register --recursive test/**/*-spec.ts",
     "tag": "bump patch -ctp README.md",
     "ts": "echo Using TypeScript && tsc --version && tsc --pretty && npm run set-bin",
@@ -38,15 +38,17 @@
     "watch:test": "npm run test -- --watch"
   },
   "dependencies": {
+    "@types/prompt-sync": "^4.1.0",
     "chalk": ">=2.3.0",
     "getopts": "^2.2.5",
     "inquirer": "^6.5.1",
     "lodash": "^4.17.15",
+    "prompt-sync": "^4.1.7",
     "prompt-sync-history": "^1.0.1",
     "version-bump-prompt": "^5.0.4"
   },
   "devDependencies": {
-    "@types/chai": "4.2.0",
+    "@types/chai": "^4.2.0",
     "@types/chalk": ">=2.2.0",
     "@types/lodash": "4.14.92",
     "@types/mocha": "2.2.46",
